feat: Email verification
Implemented email verification for account registration. Once an account is created, it recieves a verification code that will be passsed as a parameter.
A simple email is sent to the one filled out on the registration form with a link to activate the account.
Once the link is clicked by the user, their account is enabled and the confirmation code is deleted since it isn't needed anymore.

Addresses: Issue #1