feat: Implement authentication.
Created a login form with form validation
    - Checks if username or email is registered.
    - Checks if password is correct with given username or email.
    - Checks if account has been activated.

Upon a successful login, a JWT is generated that is stored in the local storage.
The JWT subject contains the user id, username, first name, last name, and the role.
The JWT subject data is then stored in a useContext hook (authState) so that the info can be used throughout multiple pages.
The variable authState also has a status property to indicate whether is user is successfully logged in or not.

For authentication, the JWT is verified to make sure it is a valid token.

Addresses: Issue #2