feat: Create delete account form
Users can now delete their own account in the settings page.
This action requires the user to type their password to confirm deleting since this is a sensitive operation.
If the account has a non-default profile picture, that image is also deleted from the Cloudinary database.
After deletion, the user is logged out and sent to the home page.
The access token is also removed from the local storage.

Address: Issue #12