feat: Finish account settings
Made the forms for the login, registration, account settings, and forgot password pages to look a little nicer.
    - Adjusted margins and title sizes.

Fixed issue where deleting the forgot password token would also delete the account.
I had to adjust the annotations in the models so that wouldn't happen.
Created a constant variable for token expiration time.
Canceling the cropping now resets the profile image to the default.

Addresses: Issues #12, #1, #2